Update cipher.py
changed chink to sink for potential future contributes (may be derogatory).

The code was deleting all user-defined functions (including load_file), constants, classes, etc. on the second and all subsequent iterations of the while loop. 
The deletion of these functions was causing the code not to work after the first iteration, as load_file function was being called but was deleted in the previous loop iteration.

Thus, I removed the global deletion code.

I was thinking to clear some variables explicitly after each iteration; however, in Python, you typically do not need to manually delete variables like this. 
Python's garbage collector will automatically free up memory from variables that are no longer in use.

You would typically only delete variables manually in Python if you're working with large amounts of data and you need to free up memory in the middle of your script. 

Please note that in the future, we may add this explicit deletions if we were to work with very large files (for now, it's more likely not necessary to manually delete variables).